rename packages to lower camel case
