+++ client/src/Services/ApiService.js
@@ -1,4 +1,4 @@
-const BASE_URL = process.env.REACT_APP_API_URL || 'http://localhost:3001';
+const BASE_URL = 'http://localhost:3001';
 
 function getPlants() {
   return fetch(BASE_URL + '/plants')
