+++ client/src/assets/cancelPlant.png
