+++ client/src/assets/greenHomepage-desktop.png
