+++ client/src/assets/yellowHomepage.png
