+++ client/src/assets/yellowHomepagenarrow.png
